refactor leave_and_join
